// apps/api/src/generated/kysely/public/PipelineInputType.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

/** Represents the enum public.PipelineInputType */
enum PipelineInputType {
  MAILGUN = 'MAILGUN',
  TWILIO = 'TWILIO',
  INBOUND_EMAIL = 'INBOUND_EMAIL',
  INBOUND_WHATSAPP = 'INBOUND_WHATSAPP',
  API = 'API',
  WEB = 'WEB',
  SLACK = 'SLACK',
  CUSTOM_EMAIL = 'CUSTOM_EMAIL',
  WEB_QUESTION = 'WEB_QUESTION',
  AIRTABLE = 'AIRTABLE',
  CHROME_EXTENSION = 'CHROME_EXTENSION',
  GMAIL = 'GMAIL',
  GRANOLA = 'GRANOLA',
  ATTIO = 'ATTIO',
  AFFINITY = 'AFFINITY',
  PIPEDRIVE = 'PIPEDRIVE',
  NATIVE_VALUATIONS = 'NATIVE_VALUATIONS',
};

export default PipelineInputType;